+++ HarvesterFront/src/app/core/models/dtos/FieldDto.ts
@@ -1,0 +1,7 @@
+export interface FieldDto {
+  name: string;
+  areaHectares: number;
+  terrainCoeff: number;
+  shapeCoeff: number;
+  cropType: string;
+}
